+++ tests/attacks/test_attack_report_formatter.py
