+++ tests/attacks/test_multiple_attacks.py
@@ -30,7 +30,7 @@ from aisdc.attacks.multiple_attacks import (  # pylint: disable = import-error
     ConfigFile,
     MultipleAttacks,
 )
-from tests.test_attacks_via_safemodel import get_target
+from tests.attacks.common import get_target
 
 # pylint: disable = duplicate-code
