+++ tests/preprocessing/test_loaders.py
@@ -17,7 +17,7 @@ from aisdc.preprocessing import loaders
 from aisdc.preprocessing.loaders import DataNotAvailable, UnknownDataset
 
 PROJECT_ROOT_FOLDER = os.path.dirname(os.path.dirname(__file__))
-DATA_FOLDER = os.path.join(PROJECT_ROOT_FOLDER, "tests", "datasets")
+DATA_FOLDER = os.path.join(PROJECT_ROOT_FOLDER, "datasets")
 
 datasets = (
     "mimic2-iaccd",
